Fix build, move files
